removed redundant files. Moved types into index in types folder for easier access. Added more typeGuards (even if not implemented yet). Added mix plugin
