+++ src/parsers/hex.ts
@@ -1,5 +1,5 @@
 import { HEXtoRGB } from "../conversions/hex";
-import { TInput, Irgba, TFormat, Ihexa, THexStr } from "../types/colormaster";
+import { TInput, Irgba, TFormat, Ihexa, THexStr } from "../types";
 import { isHEXObject } from "../utils/typeGuards";
 
 const HEXA_RE = /^#?([\da-f])([\da-f])([\da-f])([\da-f])?$|^#?([\da-f]{2})([\da-f]{2})([\da-f]{2})?([\da-f]{2})?$/gi;
