+++ src/types/colormaster.d.ts
@@ -1,190 +1,8 @@
-import { ColorMaster } from "../colormaster";
-import { HueColors, RGBExtended } from "../enums/colors";
-
-type THexDigit = "0" | "1" | "2" | "3" | "4" | "5" | "6" | "7" | "8" | "9";
-type THexLower = "a" | "b" | "c" | "d" | "e" | "f";
-type THexUpper = "A" | "B" | "C" | "D" | "E" | "F";
-
-export type THexStr =
-  | `${THexDigit}${THexDigit}`
-  | `${THexDigit}${THexLower}`
-  | `${THexLower}${THexDigit}`
-  | `${THexDigit}${THexUpper}`
-  | `${THexUpper}${THexDigit}`
-  | `${THexLower}${THexLower}`
-  | `${THexLower}${THexUpper}`
-  | `${THexUpper}${THexLower}`
-  | `${THexUpper}${THexUpper}`
-  | THexDigit
-  | THexLower
-  | THexUpper;
-
-export interface Irgb {
-  r: number;
-  g: number;
-  b: number;
-}
-
-export interface Ihsl {
-  h: number;
-  s: number;
-  l: number;
-}
-
-export interface Ihsv {
-  h: number;
-  s: number;
-  v: number;
-}
-
-export interface Ihex {
-  r: THexStr;
-  g: THexStr;
-  b: THexStr;
-}
-
-export interface Ixyz {
-  x: number;
-  y: number;
-  z: number;
-}
-
-export interface Ilab {
-  l: number;
-  a: number;
-  b: number;
-}
-
-export interface Ilch {
-  l: number;
-  c: number;
-  h: number;
-}
-
-export interface Irgba extends Irgb {
-  a: number;
-}
-
-export interface Ihsla extends Ihsl {
-  a: number;
-}
-
-export interface Ihsva extends Ihsv {
-  a: number;
-}
-
-export interface Ihexa extends Ihex {
-  a: string;
-}
-
-export interface Ixyza extends Ixyz {
-  a: number;
-}
-
-export interface Ilaba extends Ilab {
-  alpha: number;
-}
-
-export interface Ilcha extends Ilch {
-  a: number;
-}
-
-/**
- * Array value form when instantiating a RGBA or HSLA object
- */
-export type TNumArr = [number, number, number, number?];
-
-/**
- * Array value form when instantiating a HEXA object
- */
-export type TStrArr = [THexStr, THexStr, THexStr, THexStr?];
-
-/**
- * Possible operator for value adjustment of HEXA color space
- */
-export type TOperator = "add" | "sub";
-
-/**
- * Formatting options for output string
- */
-export interface IStringOpts {
-  alpha?: boolean;
-  precision?: Required<TNumArr>;
-}
-
-/**
- * Formatting options for accessibility function outputs
- */
-export interface IA11yOpts {
-  precision?: number;
-  percentage?: boolean;
-  ratio?: boolean;
-  bgColor?: TInput | ColorMaster;
-}
-
-/**
- * Options for determining if a color is readable
- */
-export interface IReadable extends Pick<IA11yOpts, "bgColor"> {
-  size?: "body" | "large";
-  ratio?: "minimum" | "enhanced";
-}
-
-/**
- * When using the monochromatic harmony, the user has more choices to pick from
- */
-export interface IMonochromatic {
-  effect?: TMonoEffect;
-  amount?: number;
-  type?: THarmony;
-}
-
-/**
- * Hue adjustment types for monochromatic harmony
- */
-export type TMonoEffect = "tones" | "tints" | "shades";
-
-/**
- * The different well know color harmonies
- * @see {@link https://www.tigercolor.com/color-lab/color-theory/color-theory-intro.htm}
- */
-export type THarmony =
-  | "analogous"
-  | "complementary"
-  | "split-complementary"
-  | "double-split-complementary"
-  | "triad"
-  | "rectangle"
-  | "square"
-  | "monochromatic";
-
-/**
- * Possible Channels for RGBA and HEXA color space
- */
-export type TChannel = "red" | "green" | "blue" | "alpha";
-
-/**
- * Possible Channels for HSLA color space
- */
-export type TChannelHSL = "hue" | "saturation" | "lightness" | "alpha";
-
-/**
- * Simplified type that combines all possible color space inputs
- */
-export type TInput = string | Irgb | Irgba | Ihex | Ihexa | Ihsl | Ihsla;
-
-export type TFormat = "invalid" | "name" | "rgb" | "hex" | "hsl" | "xyz" | "lab" | "lch";
-
-export type TPlugin = (CM: typeof ColorMaster) => void;
-
-export type TParser = (color: TInput | keyof typeof RGBExtended) => [Irgba, TFormat];
+import { IStringOpts, Ihsla, Ihexa } from ".";
+import { ColorMaster } from "..";
+import { HueColors } from "../enums/colors";
 
 export interface IColorMaster {
-  /**
-   * the RGBA color instance as an array
-   */
-  readonly array: Required<TNumArr>;
-
   /**
    * the color format of a RGBA color instance
    * @note Useful when you are using variables and are not sure what color space the color is from
@@ -368,14 +186,4 @@ export interface IColorMaster {
    * @returns The instance that was acted upon → for function chaining
    */
   rotate(value: number): ColorMaster;
-
-  /**
-   * Mix current color instance with another based on a given ratio (done in LCHA space for best results)
-   * @note The ratio represents a percentage of the input `color` that will be mixed
-   *       with the remaining portion of the current color instance
-   * @example mix('#fff', 0.2) → "mix 80% of the current color with 20% of white"
-   * @see {@link https://math.stackexchange.com/a/3263100} For argument of why LCH space should be used over LAB space
-   * @returns A new color instance corresponding to the new mixture
-   */
-  mix(color: TInput, ratio: number): ColorMaster;
 }
