+++ tests/colormaster.spec.ts
@@ -1,5 +1,5 @@
-import CM, { ColorMaster, random } from "../src/colormaster";
-import { TNumArr } from "../src/types/colormaster";
+import CM, { ColorMaster, random } from "../src/index";
+import { TNumArr } from "../src/types/index";
 
 let cm: ColorMaster;
 const TEST_COLOR = "rgba(200, 150, 100, 0.7)";
@@ -85,7 +85,6 @@ describe("darkerBy", () => {
 });
 
 describe("getters", () => {
-  test("array", () => expect(cm.array).toEqual([200, 150, 100, 0.7]));
   test("format", () => expect(cm.format).toBe("rgb"));
   test("format - invalid", () => expect(CM("magic").format).toBe("invalid"));
   test("format - name", () => expect(CM("red").format).toBe("invalid")); // name plugin is not defined
@@ -115,21 +114,6 @@ describe("rotate", () => {
   test("> 0 & < 360", () => expect(cm.rotate(180).stringHSL({ precision })).toBe("hsla(210, 47.6%, 58.8%, 0.7)"));
   test("= 360", () => expect(cm.rotate(360).stringHSL({ precision })).toBe("hsla(30, 47.6%, 58.8%, 0.7)"));
   test("> 360", () => expect(cm.rotate(361).stringHSL({ precision })).toBe("hsla(31, 47.6%, 58.8%, 0.7)"));
-});
-
-describe("mix", () => {
-  test("opposite", () => expect(CM("#FFFF").mix("#000F").stringHEX()).toBe("#777777FF"));
-  test("opposite with different opacity", () => expect(CM("#FFFA").mix("#000B").stringHEX()).toBe("#777777B3"));
-  test("red & yellow → orange", () => expect(CM("#F00").mix("#FF0").stringHEX()).toBe("#FFA000FF")); // #FFA500FF (perfect)
-  test("yellow & blue → green", () => expect(CM("#FF0").mix("#00F").stringHEX()).toBe("#00C2DEFF")); // #008000FF (perfect)
-  test("blue & red → magenta", () => expect(CM("#00F").mix("#F00").stringHEX()).toBe("#008240FF")); // #FF00FFFF (perfect)
-  test("green & magenta → blue", () => expect(CM("#008000FF").mix("#F0F").stringHEX()).toBe("#009AF8FF")); // #0000FFFF (perfect)
-  test("magenta & orange → red", () => expect(CM("#F0F").mix("#FFA500").stringHEX()).toBe("#00CBDCFF")); // #FF0000FF (perfect)
-  test("orange & green → yellow", () => expect(CM("#FFA500").mix("#008000FF").stringHEX()).toBe("#8F9A00FF")); // #FFFF00FF (perfect)
-  test("ratio < 0", () => expect(CM("#F00").mix("#ABC", -1).stringHEX()).toBe("#FF0000FF"));
-  test("ratio = 0", () => expect(CM("#ABC").mix("#F00F", 1).stringHEX()).toBe("#FF0000FF"));
-  test("ratio = 1", () => expect(CM("#F00").mix("#ABC", 0).stringHEX()).toBe("#FF0000FF"));
-  test("ratio > 1", () => expect(CM("#ABC").mix("#F00F", 2).stringHEX()).toBe("#FF0000FF"));
 });
 
 test("random generation", () => {
