+++ tests/plugins/name.spec.ts
@@ -1,4 +1,4 @@
-import CM, { ColorMaster } from "../../src/colormaster";
+import CM, { ColorMaster } from "../../src/index";
 import { extendPlugins } from "../../src";
 import NamePlugin from "../../src/plugins/name";
